+++ src/main/java/com/dnd/gongmuin/chat/dto/request/ChatMessageRequest.java
@@ -1,4 +1,4 @@
-package com.dnd.gongmuin.chat.dto;
+package com.dnd.gongmuin.chat.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
